Project3/requirements.txt
Flask==2.3.2
opencv-python>=4.8.1.78
Pillow>=10.3.0
numpy>=1.24.0
Werkzeug>=3.0.3
